trying to sort out best project architecture for type definitions updated readme
