+++ index.ts
@@ -1,7 +1,5 @@
-import transform from "./lib/transform";
-import validate from "./lib/validate";
+import libTransform from "./lib/transform";
+import libValidate from "./lib/validate";
 
-export default {
-  transform,
-  validate
-};
+export const transform = libTransform;
+export const validate = libValidate;
