+++ package.json
@@ -9,10 +9,10 @@
   "module": "./dist/objectro.esm.js",
   "types": "./index.d.ts",
   "scripts": {
-    "build": "rollup -c",
+    "build": "tsc && rollup -c",
     "lint": "eslint 'index.ts' 'lib/**/*.ts' -c .eslintrc.json --ignore-path .lintignore",
     "lint:fix": "eslint 'index.ts' 'lib/**/*.ts' -c .eslintrc.json --ignore-path .lintignore --fix",
-    "test": "jest",
+    "test": "npm run lint && jest",
     "test:watch": "jest --watch",
     "test:bench": "barbell",
     "publish": "npm run build && npm run test && npm publish"
@@ -56,6 +56,7 @@
     "jest-cli": "^24.5.0",
     "joi": "^14.3.0",
     "object-mapper": "^5.0.0",
+    "prettier": "^1.16.4",
     "prop-types": "^15.6.2",
     "rename-extension": "^3.0.0",
     "rollup": "^1.9.0",
