+++ tsconfig.json
@@ -11,6 +11,7 @@
     "module": "esnext",
     "moduleResolution": "node",
     "resolveJsonModule": true,
+    "emitDeclarationOnly": true,
     "sourceMap": true
   },
   "include": [
